Extract shared source extraction helpers into source.go
