+++ internal/cli/extract_cmd.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"os"
-	"os/exec"
 	"strings"
 )
 
@@ -34,16 +32,8 @@ func (c *CLI) runExtractCmd(ctx context.Context, args []string) error {
 	}
 	binPath := binArgs[0]
 
-	cmd := exec.CommandContext(ctx, binPath)
-	cmd.Env = append(os.Environ(), "TOBARI_EXTRACT_SOURCES="+*output)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to run %s: %w", binPath, err)
-	}
-
-	if _, err := os.Stat(*output); err != nil {
-		return fmt.Errorf("output file was not created: %s", *output)
+	if err := extractSourcesFromBinary(ctx, binPath, *output); err != nil {
+		return err
 	}
 
 	if _, err := fmt.Fprintf(c.stdout, "extracted to %s\n", *output); err != nil {
